+++ lib/cupertino_will_pop_scope.dart
@@ -1,3 +1,4 @@
 library cupertino_will_pop_scope;
 
+export 'src/conditional_will_pop_scope.dart';
 export 'src/cupertino_will_pop_scope_page_transitions_builder.dart';