+++ lib/src/cupertino_will_pop_scope_page_transitions_builder.dart
@@ -2,8 +2,7 @@ import 'package:flutter/material.dart';
 
 import 'cupertino_page_route.dart';
 
-class CupertinoWillPopScopePageTransionsBuilder
-    extends PageTransitionsBuilder {
+class CupertinoWillPopScopePageTransionsBuilder extends PageTransitionsBuilder {
   const CupertinoWillPopScopePageTransionsBuilder();
 
   @override